requirements.txt
asgiref==3.6.0
Django==3.2
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.3
Pyrebase4==4.6.0
